requirements.txt
python-dotenv>=1.0.0
pytest>=7.0.0
google-genai>=1.49.0
aiohttp>=3.8.0
google-adk>=1.18.0
aisqlite>=0.19.0